Introduced build constraints
Build constraints have been introduced in order to make the development of OS-specific features easier.
